fix(ci): correct environment variable access in run blocks (#353)
* fix(ci): correct environment variable access in run blocks

Environment variables set via GITHUB_ENV (like CMAKE_TOOLCHAIN_FILE and
VCPKG_TARGET_TRIPLET) are not immediately available in the Actions
context syntax ${{ env.VAR }}. They must be accessed as shell variables
($VAR) within 'run' blocks. This fix ensures CMake correctly picks up
the vcpkg toolchain and modern compilers, resolving Boost version
mismatch issues.

* chore: apply code formatting and fix compiler warnings

- Fixed sign-conversion and narrow conversion warnings in line_framer.cc and echo.cc.
- Resolved unused variables and handled nodiscard return values in test files.
- Improved verify.sh to automatically detect platform presets and use vcpkg.
- Applied clang-format and cmake-format to the codebase.

* fix(ci): add pull_request trigger to test and build workflows

+++ examples/serial/async/echo.cc
@@ -30,7 +30,7 @@ int main(int argc, char* argv[]) {
   uint32_t baud = 115200;
 
   if (argc > 1) device = argv[1];
-  if (argc > 2) baud = std::stoul(argv[2]);
+  if (argc > 2) baud = static_cast<uint32_t>(std::stoul(argv[2]));
 
   std::cout << "--- Async Serial Echo ---\n";
   std::cout << "Opening " << device << " at " << baud << " baud (non-blocking)\n";
